requirements.txt
alembic
marshmallow
mysql-connector-python
requests
schedule
SQLAlchemy
urllib3
vk-api